tour 2 complete

+++ View.py
@@ -190,8 +190,8 @@ class Views:
         match_list = list(tour)
         for match in match_list:
             print(f"Entrez le score pour ce match \n {match}")
-            score_list.append(int(input()))
-            score_list.append(int(input()))
+            score_list.append(float(input()))
+            score_list.append(float(input()))
         return score_list
 
     def players_modify_rank(self, players):
